requirements.txt
absl-py==2.3.0
accelerate==1.7.0
addict==2.4.0
asttokens==3.0.0
bitsandbytes==0.46.0
certifi==2025.4.26
cfgv==3.4.0
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
distlib==0.3.9
einops==0.8.1
executing==2.2.0
facexlib==0.3.0
filelock==3.18.0
filterpy==1.4.5
fonttools==4.58.1
fsspec==2025.5.1
ftfy==6.3.1
future==1.0.0
grpcio==1.72.1
huggingface-hub==0.32.3
icecream==2.1.4
identify==2.6.12
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
Jinja2==3.1.6
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.44.0
lmdb==1.6.2
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
nodeenv==1.9.1
numba==0.61.2
numpy==2.2.6
openai-clip==1.0.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
protobuf==6.31.1
psutil==7.0.0
Pygments==2.19.1
pyiqa==0.1.13
pyparsing==3.2.3
pytest==8.4.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.3
sentencepiece==0.2.0
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tifffile==2025.6.1
timm==1.0.15
tokenizers==0.15.2
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchvision==0.22.1+cu128
tqdm==4.67.1
transformers==4.37.2
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
virtualenv==20.31.2
wcwidth==0.2.13
Werkzeug==3.1.3
yapf==0.43.0
